Add configuration and acceptance documentation for parameter linkage and migration
- Introduced a new document detailing user operation methods for configuration templates, emphasizing the use of `hbm.cfg` and `lpddr.cfg`.
- Documented three categories of parameters and their usage principles, including independent control inputs and derived values.
- Established rules for schema 3, including capacity calculations and timing linkage boundaries.
- Added a comprehensive acceptance record for configuration refactoring, including backup procedures and implementation results.
- Implemented automatic and external validation checks, confirming successful test outcomes across various configurations.
- Updated output formats to include JSON schema for better machine readability and diagnostics.
- Revised existing documentation to reflect changes in parameter handling and output reporting.

+++ include/hbm_sim/dram/spec.hpp
@@ -316,7 +316,9 @@ struct DramSpec {
   std::string vendor_profile = "generic";
   std::string mode_profile = "default";
   int speed_bin_mbps = 0;
-  int density_gb = 0;
+  // Gibit per die (HBM) or per subchannel/rank (LPDDR). Fractional research
+  // geometries must not be rounded to an unrelated density table row.
+  double density_gb = 0;
   int stack_height = 0;
   // true 时，tick() 每周期分别尝试一条 column 和一条 row 命令。
   bool dual_command_bus = false;
